requirements/requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
celery==5.3.4
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.3
Django==3.2
django-utils2==3.0.0
djangorestframework==3.14.0
drf-spectacular==0.26.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.7
pyrsistent==0.19.3
python-dotenv==1.0.0
python-utils==3.5.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.15
django-celery-beat==2.5.0
drf-generators==0.5.0
pyecharts==2.0.4
pytest==7.4.2
pytest-django==4.5.2